sopel: move deprecated() to sopel.deprecated
The `deprecated()` function is used everywhere, whenever we need to
deprecate something. Sometimes, that somethings is right where this
function is defined: in `sopel.tools`, causing cyclic import errors.

In an effort to reduce cyclic import errors, and to organize the code a
bit better (mostly by removing everything from
`sopel/tools/__init__.py`), this function is moved to its own sopel's
submodule: `sopel.deprecated`.

Documentation has been updated accordingly, and every part of Sopel that
used `sopel.tools.deprecated` is now using
`sopel.deprecated.deprecated`.

In order to prevent yet another cyclic import error, `deprecated` no
longer uses `sopel.tools.get_logger` to get the logger for an external
plugins (that isn't using the namespace), but tries to mimic its
behavior for namespace plugins only.
